PmergeMe has beenoverhauled with the full sorting functionality added.
Removed non-const versions of expand and shrink as they modified the original object instead of spawning new objects during recursive calls.

Due to the templated nature of the sorting algorithm, some undefined behavior may be encountered.

Requires a bit of additional testing with sequences of different sizes and later double-checking that the container has been sorted.